make FastpathWorker
all fastpaths (agent output, substrate ingress) now share same worker

drop_and_count is now part of this worker
